Initial commit
First pass at a new initrd implementation.  Right now, it's a very simple
/init and we use nash's switchroot command via switch_root.  The idea is to
get a switchroot binary into util-linux that can be used instead

generate.sh will let you generate an initrd image.  Note that the generator
is intentionally super simple right now and is an area that will need a lot
of work once we're happier with how the /init process runs