We now know how to configure the network interfaces.
Provided that you are configuring them via DHCP. RARP, BOOTP, and static
configuration are not written yet.

Also, adding nic drivers really bloats the initrd.

I am looking for feedback on these patches from people who actually implement
booting over the network.  This patch series does not include support for that
yet -- you will not find nfsroot= handling, or booting to a fs supplied by
dhcp yet.  I do want to make it as easy as possible to add support for booting
over the network, as well as making it easy for people to customize things to
meet their site requirements.

This patch series is also available as the network-configurability branch at
http://git.fnordovax.org/dracut.  It may be rebased without warning to keep
it in sync with the rest of dracut.